+++ packages/devtools/src/formats/runtime.ts
@@ -338,21 +338,26 @@ export function createRuntimeDefinition(
     Object.assign(runtime.objects, builtModel.objects)
     Object.assign(runtime.enums, builtModel.enums)
     // Attach entry-point to account store based on relation type
-    if (modelDefinition.accountRelation != null) {
-      const key = camelCase(modelName)
-      if (modelDefinition.version !== '1.0' && modelDefinition.interface) {
-        runtime.accountData[key + 'List'] = { type: 'connection', name: modelName }
-      } else {
-        const relationType = modelDefinition.accountRelation.type
-        if (relationType === 'single') {
-          runtime.accountData[key] = { type: 'node', name: modelName }
-          // @ts-ignore TS2367, should be unnecessary check based on type definition but more types
-          // could be added later
-        } else if (relationType === 'list') {
+    const key = camelCase(modelName)
+    if (modelDefinition.version !== '1.0' && modelDefinition.interface) {
+      runtime.accountData[key + 'List'] = { type: 'connection', name: modelName }
+    } else {
+      const relationType = modelDefinition.accountRelation.type
+      switch (relationType) {
+        case 'list':
           runtime.accountData[key + 'List'] = { type: 'connection', name: modelName }
-        } else {
-          throw new Error(`Unsupported account relation type: ${relationType as string}`)
-        }
+          break
+        case 'set':
+          // relation to single document in the set based on input
+          runtime.accountData[key] = { type: 'set', name: modelName }
+          // relation to all documents in the set
+          runtime.accountData[key + 'List'] = { type: 'connection', name: modelName }
+          break
+        case 'single':
+          runtime.accountData[key] = { type: 'node', name: modelName }
+          break
+        default:
+          throw new Error(`Unsupported account relation type: ${relationType}`)
       }
     }
   }