+++ packages/server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @composedb/server
 
+## 0.7.0
+
+### Minor Changes
+
+- Add support for models using the new `set` account relation
+- Add support for the `shouldIndex` metadata flag
+- Add support for immutable fields
+- Add support for optional relation fields
+
 ## 0.6.0
 
 ### Minor Changes